improved code a bit and added first writing binary function
commiting to have a backup.
going to change the analyse_operands() to be compatible with parameteres with space in between.

next commit should be (if there is nothing i need to save in the middle) a working first-pass, with good operation, labels, and parameters detection and perfect binary writing, need just to fix some error catching and other small things